test(civitai-browser): fix order-dependent dispatcher pump flakiness
RunWithDispatcherPumpAsync polled Dispatcher.UIThread.RunJobs() from a
background thread hoping something would eventually wake it, but this
project never initializes an Avalonia platform, so Dispatcher.UIThread
falls back to Avalonia's NullDispatcherImpl: Signal()/Signaled are
no-ops, so a queued operation only gets serviced when *some* thread,
anywhere in the process, happens to call RunJobs() again. In the full
suite there's always ambient pumping pressure from unrelated
concurrently-running tests; filtered down to just the CivitaiBrowser
classes that pressure collapses and the fixed 5s timeout occasionally
loses the race (confirmed live: 318 pump iterations, no throw, no
progress).

Move the six tests that actually await Dispatcher.UIThread.InvokeAsync
mid-flight into DiffusionNexus.IntegrationTests as [AvaloniaFact]
tests, which already carry a real headless Avalonia dispatcher with
genuine signaling and parallelization disabled — a plain await
resolves them with no manual pump needed (verified 10/10 via spike
before committing to the move). Delete the now-unused
RunWithDispatcherPumpAsync helper from both original files.
